# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Second Brain - academic argument analysis tool
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type SessionInputMethod =
  (typeof SessionInputMethod)[keyof typeof SessionInputMethod];

export const SessionInputMethod = {
  paste: "paste",
  blank: "blank",
  upload: "upload",
} as const;

export type SessionStatus = (typeof SessionStatus)[keyof typeof SessionStatus];

export const SessionStatus = {
  draft: "draft",
  analyzing: "analyzing",
  complete: "complete",
} as const;

export interface Session {
  id: number;
  title: string;
  content: string;
  inputMethod: SessionInputMethod;
  status: SessionStatus;
  /** @nullable */
  rating: number | null;
  analysisCount: number;
  createdAt: string;
  updatedAt: string;
}

export type SessionSummaryStatus =
  (typeof SessionSummaryStatus)[keyof typeof SessionSummaryStatus];

export const SessionSummaryStatus = {
  draft: "draft",
  analyzing: "analyzing",
  complete: "complete",
} as const;

export interface SessionSummary {
  id: number;
  title: string;
  status: SessionSummaryStatus;
  /** @nullable */
  rating: number | null;
  analysisCount: number;
  biasCount: number;
  fallacyCount: number;
  oppositionCount: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateSessionBodyInputMethod =
  (typeof CreateSessionBodyInputMethod)[keyof typeof CreateSessionBodyInputMethod];

export const CreateSessionBodyInputMethod = {
  paste: "paste",
  blank: "blank",
  upload: "upload",
} as const;

export interface CreateSessionBody {
  title: string;
  inputMethod: CreateSessionBodyInputMethod;
  content?: string;
}

export interface UpdateSessionBody {
  title?: string;
  content?: string;
}

export interface RateSessionBody {
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
}

export type AnalysisItemType =
  (typeof AnalysisItemType)[keyof typeof AnalysisItemType];

export const AnalysisItemType = {
  bias: "bias",
  fallacy: "fallacy",
  opposition: "opposition",
} as const;

export interface AnalysisItem {
  id: number;
  sessionId: number;
  type: AnalysisItemType;
  content: string;
  explanation: string;
  /** @nullable */
  accepted: boolean | null;
  createdAt: string;
}

export interface Analysis {
  sessionId: number;
  items: AnalysisItem[];
  biasCount: number;
  fallacyCount: number;
  oppositionCount: number;
  acceptedCount: number;
}

export interface UpdateAnalysisItemBody {
  accepted: boolean;
}

export interface ProgressDataPoint {
  sessionId: number;
  date: string;
  totalItems: number;
  acceptedItems: number;
  /** @nullable */
  rating: number | null;
  title: string;
}

export interface ProgressSummary {
  totalSessions: number;
  completedSessions: number;
  /** @nullable */
  averageRating: number | null;
  totalBiasIdentified: number;
  totalFallaciesIdentified: number;
  totalOppositionGenerated: number;
  totalItemsAccepted: number;
  dataPoints: ProgressDataPoint[];
}