# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Second Brain - academic argument analysis tool
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sessions/{id}/analysis:
    get:
      operationId: getAnalysis
      tags: [analysis]
      summary: Get analysis results for a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "404":
          description: Analysis not yet run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sessions/{id}/analysis/items/{itemId}:
    patch:
      operationId: updateAnalysisItem
      tags: [analysis]
      summary: Accept or reject an analysis item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAnalysisItemBody"
      responses:
        "200":
          description: Updated analysis item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    AnalysisItem:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        type:
          type: string
          enum: [bias, fallacy, opposition]
        content:
          type: string
        explanation:
          type: string
        accepted:
          type: ["boolean", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - type
        - content
        - explanation
        - accepted
        - createdAt
    Analysis:
      type: object
      properties:
        sessionId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/AnalysisItem"
        biasCount:
          type: integer
        fallacyCount:
          type: integer
        oppositionCount:
          type: integer
        acceptedCount:
          type: integer
      required:
        - sessionId
        - items
        - biasCount
        - fallacyCount
        - oppositionCount
        - acceptedCount
    UpdateAnalysisItemBody:
      type: object
      properties:
        accepted:
          type: boolean
      required:
        - accepted
...