# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Second Brain - academic argument analysis tool
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all sessions
 */
export const ListSessionsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  status: zod.enum(["draft", "analyzing", "complete"]),
  rating: zod.number().nullable(),
  analysisCount: zod.number(),
  biasCount: zod.number(),
  fallacyCount: zod.number(),
  oppositionCount: zod.number(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListSessionsResponse = zod.array(ListSessionsResponseItem);

/**
 * @summary Create a new session
 */
export const CreateSessionBody = zod.object({
  title: zod.string(),
  inputMethod: zod.enum(["paste", "blank", "upload"]),
  content: zod.string().optional(),
});

/**
 * @summary Get a session by ID
 */
export const GetSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetSessionResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  inputMethod: zod.enum(["paste", "blank", "upload"]),
  status: zod.enum(["draft", "analyzing", "complete"]),
  rating: zod.number().nullable(),
  analysisCount: zod.number(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a session's content or title
 */
export const UpdateSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateSessionBody = zod.object({
  title: zod.string().optional(),
  content: zod.string().optional(),
});

export const UpdateSessionResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  inputMethod: zod.enum(["paste", "blank", "upload"]),
  status: zod.enum(["draft", "analyzing", "complete"]),
  rating: zod.number().nullable(),
  analysisCount: zod.number(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a session
 */
export const DeleteSessionParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get analysis results for a session
 */
export const GetAnalysisParams = zod.object({
  id: zod.coerce.number(),
});

export const GetAnalysisResponse = zod.object({
  sessionId: zod.number(),
  items: zod.array(
    zod.object({
      id: zod.number(),
      sessionId: zod.number(),
      type: zod.enum(["bias", "fallacy", "opposition"]),
      content: zod.string(),
      explanation: zod.string(),
      accepted: zod.boolean().nullable(),
      createdAt: zod.coerce.date(),
    }),
  ),
  biasCount: zod.number(),
  fallacyCount: zod.number(),
  oppositionCount: zod.number(),
  acceptedCount: zod.number(),
});

/**
 * @summary Accept or reject an analysis item
 */
export const UpdateAnalysisItemParams = zod.object({
  id: zod.coerce.number(),
  itemId: zod.coerce.number(),
});

export const UpdateAnalysisItemBody = zod.object({
  accepted: zod.boolean(),
});

export const UpdateAnalysisItemResponse = zod.object({
  id: zod.number(),
  sessionId: zod.number(),
  type: zod.enum(["bias", "fallacy", "opposition"]),
  content: zod.string(),
  explanation: zod.string(),
  accepted: zod.boolean().nullable(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Submit a session rating
 */
export const RateSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const rateSessionBodyRatingMax = 5;

export const RateSessionBody = zod.object({
  rating: zod.number().min(1).max(rateSessionBodyRatingMax),
});

export const RateSessionResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  inputMethod: zod.enum(["paste", "blank", "upload"]),
  status: zod.enum(["draft", "analyzing", "complete"]),
  rating: zod.number().nullable(),
  analysisCount: zod.number(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Get overall progress summary
 */
export const GetProgressResponse = zod.object({
  totalSessions: zod.number(),
  completedSessions: zod.number(),
  averageRating: zod.number().nullable(),
  totalBiasIdentified: zod.number(),
  totalFallaciesIdentified: zod.number(),
  totalOppositionGenerated: zod.number(),
  totalItemsAccepted: zod.number(),
  dataPoints: zod.array(
    zod.object({
      sessionId: zod.number(),
      date: zod.coerce.date(),
      totalItems: zod.number(),
      acceptedItems: zod.number(),
      rating: zod.number().nullable(),
      title: zod.string(),
    }),
  ),
});