# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Second Brain - academic argument analysis tool
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisItem } from "./analysisItem";

export interface Analysis {
  sessionId: number;
  items: AnalysisItem[];
  biasCount: number;
  fallacyCount: number;
  oppositionCount: number;
  acceptedCount: number;
}