# lib/api-zod/src/generated/types/analysisItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Second Brain - academic argument analysis tool
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisItemType } from "./analysisItemType";

export interface AnalysisItem {
  id: number;
  sessionId: number;
  type: AnalysisItemType;
  content: string;
  explanation: string;
  /** @nullable */
  accepted: boolean | null;
  createdAt: Date;
}