# lib/api-zod/src/generated/types/updateAnalysisItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Second Brain - academic argument analysis tool
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateAnalysisItemBody {
  accepted: boolean;
}